Ensure created project gets deleted when all jobs run to completion
The `--delete-when-idle=DURATION` flag instructs the ci-operator to
create a pod that will delete the namespace if there are no running,
pending, or unknown restartPolicy=Never pods for longer than DURATION.
This ensures that the namespace is properly cleaned up when jobs exit.
The created pod has a hardcoded 12 hour limit before the namespace is
torn down automatically.