Allow templates to be used as execution steps
Templates are passed to the command, parsed, and then used as arbitrary
execution units in the CI graph. The operator requires that each
template contain at least one restart=Never pod and will set up
dependencies on earlier stages based on the parameters passed as input.
The following parameters are automatic:

* JOB_NAME: the name of the job from the Job spec
* JOB_NAME_HASH: a short hash of the job name
* JOB_NAME_SAFE: the job name converted to a Kubernetes safe name
* NAMESPACE: the target namespace
* IMAGE_FORMAT: requires the release stage be defined
* RPM_REPO: a URL pointing to the hosted RPM repo
* IMAGE_*: depends on the release stage and places the wildcard value
    into IMAGE_FORMAT as ${component}

Environment variables can underfill the job.

The pods in each template are considered successful when they reach
phase=Succeeded and considered failed if any container or init container
has a non-zero exit code. The template name is used as the unique key.