Include all dependencies of the build in the input hash
The hash used for the namespace name needs to include the external
dependencies we pull for images as well as the build configuration and
source code in order to ensure we have reproducible builds.

Add a new Step method `Inputs(...) api.InputDefinition` which returns an
opaque list of inputs that can be combined into the final hash. After
loading the steps we resolve all inputs and calculate the hash for the
inputs, then use that in the input name.

Because namespace is an input to the steps but can only be determined
after we resolve inputs defined in other namespaces, we must switch
steps to lazily include their namespace (so we pass cluster scoped
clients to steps instead of namespace scoped steps).

This makes the ci-operator hermetic with respect to its inputs.