Gather artifacts from test pods and template pods
A test pod must have "artifact_dir" defined and the user must pass a top
level --artifact-dir flag to gather artifacts. A template pod will
gather artifacts if the flag is set and the pod has a volume called
"artifacts" and that volume is mounted into any of the pod containers.

Artifacts are gathered by invoking a tar exec session into a busybox
sidecar. After all containers have gathered logs a file is removed in
the container via exec which cases the session to exit.