Multiple fixes to artifacts in edge cases
Don't declare the job finished until all containers have at least
started. Don't include the "artifacts" container in the list to wait
for. Properly hold exiting artifact upload for a pod until the final
step is complete (removing the marker file) which avoids races for the
pods. For pods in templates, check for artifact containers after the
template instance is created in order to handle parameter resolution.
Finally, if the pod is already running in a template clean it up before
continuing.