Support accessing input images from another cluster
The `cluster` field may be set on input images and the release tag
configuration to anonymously access image streams located on a public
cluster (the image stream, tags, images, and pulls must all be available
anonymously). This allows a repo-specific config file to be checked in
that can be overriden during CI by the job executor.