Add a promotion definition that will copy pipeline tags out
Branch jobs should be able to promote the known tags to a target
location when needed. The alternative is to duplicate the list of images
known in the config, which complicates setup. --promote is an optional
flag that instructs the ci-operator to also attempt a promotion. This
prevents the job from being able to automatically promote (when in a
repo config). The promotion step is executed after the graph and
promotes anything that was successfully built.