When pods and builds fail, report more information about them
Some jobs were being OOMKilled, but the data was not being reported back
so it wasn't obvious that the container needed memory adjustments. For
both builds and pods, try to capture the reason the pod failed and also
include termination message. Use terminationMessagePath fallback to logs
on pods we run in order to capture their last few lines.

Write the extended error output to the junit report if captured.