Perform an additional check to see if src and dst are the same
When running inside a cluster, the dst client hostname will be the kubernetes service IP, but the src client hostname might be the external URL to the cluster. Perform a check on the input by seeing if is reachable by the dst client and the target has the same UID.

Allows us to bypass image stream import for most clusters.