Add a release creation step
The step reuses the pod test logic (factored out) to run a cli command
to craft a new release from the latest stable stream in a PR. The release
payload is published to "release:latest" in the image stream. The step
is always run as long as there is a release tag specification. The step
does nothing if no image "cluster-version-operator" is present in the stable
image stream.

To reference, templates may use the RELEASE_IMAGE_LATEST env var.