Refactor generator to write at the directory level
We want to enforce that jobs are sharded by the branch and the job type,
which we can only do if we load and write at the directory level,
instead of the single file level. This patch reconfigured the generator
and the determinizer to operate at the directory level and to shard jobs
by branch and by job type.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>