Honor specific fields in Prow configuration
There exist fields in the Prow Job configuation that are only germane to
the Prow triggering and scheduling algorithms. We cannot have the source
of truth for these fields living in the CI Operator configuration, but
we can set defaults for these fields using the generator. While the
defaults will be generally applicable, we cannot ensure that they will
be valid in all cases. This patch allows changes to those fields to be
honored by the generator to allow users that opt into that level of
configuration complexity leeway when changing job configuration.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>