Substitute targets that are completely covered by input ENV
A target that defines one or more provides that are all covered by
environment variables defined on the process should be substituted
by a shim job that prints no output. This makes skipping certain
execution much simpler and allows parts of the graph to be bypassed
when no build is necessary.