When generating the postsubmit images job, only promote on exact matches
Because prowjob branches are regexps, the "master" branch matches
`.*master.*`. If a team has a working branch like `master-dockerless`
and merges to that branch, the promote images job will stomp the images
that the `master` branch promotes.

With this change, the image promotion job is set to exactly match the
requested branch, unless the branch looks like a regex (has characters
outside the normal branch name set). This prevents accidental image
promotion that occurs on feature branches.