Fix a possible deadlock in waitForCompletionOrTimeout
There is a small window between first checking the Pod status on
L586/L590 and setting up the watcher on L594. If the Pod status changes
exactly in this window, we may end up waiting indefinitely on L604 for
an event that already happened.

Although this is very unlikely to happen in reality, I was seeing it
happen occasionally when unit-testing with fakes.