Unit tests for `test.go` step
These tests finally make use of the cluster mocking libraries. The mock
Pods.Create() method needed to be extended to create Pod with
Status.Phase=Pending, because the libraries do not put any defaults into
the fields and ci-operator code was failing on Phase="".

The `executeStep` method can now be provided a function that is executed
in a goroutine before the `Run()` method is executed, and which can be
used to simulate the expected fake cluster behavior.