Fix more deadlock opportunities
Previously, these were all using the following pattern:

1. Check for the current status of $RESOURCE and exit on $CONDITION
2. Set up a watcher on $RESOURCE
3. Loop until event on $RESOURCE happens and $CONDITION holds, then exit

The problem is that if the event to make $CONDITION hold happens exactly
between 1. and 2. (after we check first but before we set up the
watcher), we miss it and hang in the for loop waiting.

This is preventing by setting up the watcher first, so it receives
events even while we check the status initially.