Add an optional --input-hash flag for making namespaces unique
Some jobs vary only by job id (when using deterministic external inputs)
and are valuable to run in parallel, such as post submit soak tests where
we want to run the same e2e input multiple times in a row. Allow the caller
to pass one or more input values to add to the input hash to ensure the
created namespace is unique. If the user specifies --namespace and doesn't
pass {id} as part of that value, this flag has no effect.