The build api reports failure before log snippet
Work around a bug/unfriendly API in builds by waiting up to an
arbitrary amount of time between when the build fails and when we
think a logSnippet might grace us with its presence. This ensures
that when docker builds fail in ci-operator we report the snippet
in the junit output so that a human gets a nice message.