Unify how rehearse works with paths inside openshift/release
Previously, path detection was inconsistent between modules. Diff module
assumed the tool is run in a working directory with a working copy of
`openshift/release` to be rehearsed, and hardcoded the necessary paths
inside `openshift/release`. Rehearsal code accepted the full path to
ci-operator configs as a parameter to the binary.

This was unified so that the binary accepts a path to
`openshift/release` working copy. Current working directory is the
default value. All code that needs to read something from it hardcode
relative paths as constants.