Allow pods to declare that they want one junit per container, instead of pod
Setting the annotation `ci-operator.openshift.io/container-sub-tests` to
a list of comma delimited container names will replace the overall junit
result reported for that pod or template with one junit result per test.

This allows a pod that has both setup and test containers to get per task
junits which aids debugging.

Also simplify the names of the steps as printed in junit to make them
easier to scan.