Allow images to be excluded from promotion, no matter what
Some test environments may wish to build an image and put it in the
payload for testing, but not promote that image. An example of this
is the new machine-os-content image that will be built for RHCoS -
a few projects like origin and crio need to be able to take the
existing machine-os-content as a prereq, layer on top of it their
test RPMs, and then verify the complete payload as an e2e test (i.e.
test RHCoS with their latest code). Since this image is test only,
we need a clean way to indicate it's not for promotion, but we
want to build it every time we build the payload.

This will also be simpler for folks to reason about than optional,
which was really only for artifact style images that currently
aren't ever tested or added than the payload.

In promotion config, add:

    excluded_images:
    - image-1
    - image-2
    - ...

and the referenced images will never be promoted.