Embed rehearsed job branch into rehearsal job's context
Previously, when a PR was attempting to add two jobs with same context,
but different branch, pj-rehearse would generate two rehearsal jobs with
the identical context, leading to both of them reporting into the same
GH context. The fix is to include target branch name into a context to
distinguish them.