Generate trigger regex as upstream does
We should not be matching the literal `/test all` with our triggers.
It's not best practice and should be avoided. This patch refactors our
trigger generation to use the same logic as the upstream defaulting
would, generating the identical regex for a job name as would happen if
no regex was present in the Prow config at all. We still need to
generate our own as we want the name in the regex to be a subset of the
full name of the job.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>