Force rehearsals to have an unified trigger
Rehearsals cannot be triggered directly (Prow will only kick jobs that
are configured for the given repo on /test XXX commands), so we
work around it by making it share RerunCommand with the `pj-rehearse`
job (the driver). RerunCommand is only used to give users hints in a
comment, so this just makes the bot stop suggesting comments that would
not work.