If cancelled, promotion was still being invoked
The Run method, when cancelled by a context interrupt, would not
return an error if no error had occurred during step execution.
This caused promotion to think that the run was successful, and
we attempt to promote things we haven't created yet. This could
result in an older image being promoted by accident.