Retry promotion longer
While testing full release builds I was able to cause a bunch of
builds to line up and attempt to promote at the same time (12)
which exceeded our default retry window of 5. Increase the retry
count and add a slight backoff so that even highly parallel
rebuilds have a good chance of completing.