Avoid touching files already in the src image and causing a copy up
In overlayfs, changing file attributes causes the entire file to be
copied up. When the origin master src cache image is used (which has
/go/src/%/.git prepopulated) chmod g+wx -R caused every file in the
directory to be changed, which meant the top layer for origin src
images ended up being as big as the repo, about 1gb.

This changes only directories that don't already have the correct
permissions, which is the original intent of the change.