base: Allow base_images to be defaulted from tag_specification
While testing image builds with ci-operator it was difficult to
take a bunch of ci jobs and test a complete configuration that has
a different base image or source location. One difficulty was that
we had redundant info in almost all ci-operator job definitions
from the base image that comes from the tag definition. This
change simplifies base images a bit when pulling from a common
root by allowing the cluster, namespace, and name field for
base_images and base_rpm_images to default to the tag specification.

E.g.:

```
tag_specification:
  name: '4.0'
  namespace: ocp
base_images:
  base:
    name: '4.0'
    namespace: ocp
    tag: base
```

can be reduced to

```
tag_specification:
  name: '4.0'
  namespace: ocp
base_images:
  base:
    tag: base
```