Require full branch name with --git-ref
This bit me in https://github.com/openshift/origin/pull/22081. My PR
branches are all of the form `pr/$name`, but using
`--git-ref jlebon/origin@$name` still worked because `git ls-remote`
just needs a trailing match. However, it will then throw off the image
builds in OCP.

As a sanity check, if the ref is a regular branch and not e.g. a tag, or
pull request, let's verify that the full branch name was specified.