Remove duplicate ci-operator config loading code
With the ability to compare ci-operator config we needed to implement a
structure that holds all ci-operator configs in memory. When inlining
ci-operator config to rehearsed jobs, we loaded from disk the config
file content to be inlined. This is no longer necessary, because we
already have that config in memory, so we can marshal it and inline the
result.