rehearse: optimize git manipulation when loading configs
Previously we loaded all three CI config fragments (Prow config,
ci-operator config, templates) separately. Because we need to load the
configs from two different revisions of the release repo, all three
loading methods handled the "check out master revision, load config,
check out tested revision" procedure themselves, which resulted in the
whole operation being performed three times unnecessarily. This was
optimized so that we separate loading of all config from the git
operation.

A little tricky area is the error handling and propagation. The errors
occured during config loading itself are not propagated via error values
but by the appropriate field of the config having a nil value. This is
because we need granular information about which configs loaded
correctly and which not, so the tool can decide a right course of
action. The decisions are as follows:

- Exit when any Prow config version loading failed, because we need to
  compare both versions to compute which jobs should be rehearsed.
- Exit when PR version of templates or ci-operator config loading
  failed, because we need these versions to provide them to rehearsed
  jobs.
- Do not compare templates when master version of templates fails to
  load.
- Do not compare ci-operator config when master version of these configs
  fails to load.