Default `stable` image stream from RELEASE_IMAGE_LATEST
Periodic jobs that want to run a PR from a known `RELEASE_IMAGE_LATEST`
need access to the `cli` and `installer` images out of the payload. This
fixes problems we have testing older installers and cli binaries against
newer payloads and fixes the current break in the release controllers.
This is also required for properly testing the installer.

Use standard oc commands (that are API forward compatible) to extract
the `cli` image and the `installer` image from the payload and set them
on stable if RELEASE_IMAGE_LATEST is provided as an input (as they are
for the release blocking jobs) or on stable-initial if
RELEASE_IMAGE_INITIAL is provided.

Hide nested container output because it is confusing when leveraging a
pod from within an existing step.