Implement a Golang cluster installer job that hides the template
Take the existing cluster-launch-installer-e2e template and hide it
inside a e2e step definition. Support an `upgrade` flag that alters
the parameters passed to the template such that the installer image
from the input (either release tag or an actual payload) is used to
install, then launch the upgrade tests from the job definition.

This sets the path for moving this functionality into the operator.

The user still has to upload the right cluster profile secret, but
now we print an error if we don't find it (for stronger typing).
Regular users can launch an upgrade from a single command.

Remove the previous behavior of not returning parameter links when
the env var is set - the logic in defaults.go is sufficient to
prune steps that are fully specified. The release command now
cannot be overriden by env (it imports when env is specified) and
must always be run.