Create an initial release payload as `release:initial`
When we tag images in, create a second image stream `stable-initial`
that will not be modified and then let a secondary task create an
image stream from that alternate source. If RELEASE_IMAGE_INITIAL
is already specified, skip the action and instead tag it to that
location.

Simplify the structure of the AssembleReleaseStep to handle
both initial and latest.

Makes it easy to test an upgrade from the previous state to the
current state (post change) while also allowing for injecting an
initial step.