Use UID precondition to prevent races with other ci-operators
When a /retest retries a build, we're going to race between multiple
jobs. Use precondition on UID to avoid too much of a race. We
probably need the build wait function to also handle cases like
this, but this is good enough for now (most of the racers will be
happening in the same window).