prowgen: prune stale generated jobs
Label the generated jobs so that Prowgen can recognize jobs it generated
in the past. When merging newly generated jobs into the old job
configuration, prune jobs that were generated earlier but were not
generated by the current run. This will prevent stale jobs being kept
in the release repo after the reason to generate the jobs is gone (for
example, when someone removes a clause in the `tests` array in their
ci-operator config, but forgets to remove the job that was generated for
it).