Handle existing imagestreamtags better by updating if it exists
The previous code did a delete, then a create, but in practice, the
delete can succeed, and then the create can fail because the deletion is
async. The existing logic was incorrectly assuming an AlreadyExists
error was acceptable, but in reality, it is not, as this would mean the
imagestreamtag does not match the desired output, and it may already
exist due to the deletion not being finished when the create call is
made.

To remedy this, instead attempt to create first, and if it fails because
the tag already exists, update the tag to the desired state. This avoids
waiting for delete to finish and also has the added benefit of ensuring
the tags always exist, and have no gaps of time where an image pull
could fail due the tag being deleted prior to creation as was possible
with the old implementation.