Allow config-brancher to mirror promotion outputs
When we are creating a release branch for the same version that the
development branch is using, that release branch needs to mirror the
promotion config from the development branch but in a disabled manner.
The config-brancher tool can now create these mirrors and unmirror to do
the actual bump for the development branch in the future.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>