Refactor config brancher to populate many releases
The config-brancher tool has been refactored to support the new paradigm
for branches. This tool is intended to make the process of branching and
duplicating configuration for the CI Operator easy across many
repositories.

We select a set of repositories to operate on by looking at which are
actively promoting images into a specific OpenShift release, provided by
`--current-release`. Branches of repos that actively promote to this
release are considered to be our dev branches.

Once we've chosen a set of configurations to operate on, we can do one
of two actions:
 - mirror configuration out, copying the development branch config to
   all branches for the provided `--future-release` values, not changing
   the configuration for the dev branch and making sure that the release
   branch for the version that matches that in the dev branch has a
   disabled promotion stanza to ensure only one branch feeds a release
   ImageStream
 - bump configuration files, moving the development branch to promote to
   the version in the `--bump` flag, enabling the promotion in the
   release branch that used to match the dev branch version and
   disabling promotion in the release branch that now matches the dev
   branch version.

This will support the workflow where we have a dev branch coexisting
with many release branches, one of which mirrors the dev branch
promotion configured in a disabled way.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>