Add a tool to publish merge blocking issues for releases
When we are automatically fast-forwarding branches from the development
branch for future releases, it is not valid to merge to those branches
directly. The new `blocking-issue-creator` issue posts issues to the
repositories for which we are managing future release branches in the
format that `tide` expects to block automatic merging to those branches.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>