Restore force deletion of tag in outputImageTagStep
machine-os-content promotion uses an output_image_tag_step to take
an external dependency and override the stable tag. When we reverted
the update code (for other reasons) the delete case wasn't restored.

Add a TODO that in the future we may want "force" to be a different
step or have optional behavior, since the dominant pattern should be
"if-not-exists".