Attempt to save current namespace resources to artifacts during exit
In order to better debug failure modes involving infrastructure, have
the ci-operator attempt to capture what was occurring at the time it
exited in the cluster. This also allows to audit when the operator isn't
giving us the results we expect, such as when it fails to properly
publish or promote images.