Remove extra merge blocker issues when they exist
The GitHub GraphQL API was giving us success in creating issues in
repositories that had issues disabled, but was not letting us list
issues there. Therefore, we created more than one issue at once for
those repos even though our logic would have stopped that if we were
able to see that we had already created an issue. Now, we need to clean
up the extra issues on those repos after we've enabled issues on them.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>