Rehearsal metrics: Find and count stale rehearsal occurences
Stale rehearsals happen when the rehearsal tool selects different jobs
to rehearse for a single commit (so a /retest needs to be involved).
These rehearsals are confusing when they fail, because they are very
hard to get rid of (they cannot be separately rerun, and pj-rehearse
does not attempt to re-run them either). This change makes the metrics
tool track occurrences of this.

Related to:
 - DPTP-368
 - https://github.com/openshift/ci-operator-prowgen/pull/131