The old delete then create code creates a gap where two PR jobs race
The first creates stable then continues - the second deletes the tag,
but then first tries to use the tag and fails.

Replace delete then create with update image stream tag, retry on
conflict, and if it doesn't exist (or another job races with us and
it does exist), ignore that error.

There is a brief window where if the source tag is external the image
import controller might still be importing the image, but most code
in ci-operator now at least checks that condition. If it comes up we
can have output image stream tag loop and wait for import like the
release command does.