Make artifacts container survive longer after signal
Artifacts container lives for a while after a first ci-operator instance
manages to fetch the artifacts from it. There may be *other* ci-operator
instance waiting on the pod and fetching artifacts, though, so let them
have more time to finish their artifact grab.

This is not a proper fix of the race, just a bandaid to reduce
occurrences in production.

See: https://bugzilla.redhat.com/show_bug.cgi?id=1693865#c12