prowgen: Remove some dead code
Previously we needed to handle labels specifically in postsubmits
because we were marking them with `artifacts: images`. We no longer do
that, but we were still creating an empty map, passing it through and
API that was more complicated than needed and copying it around. All
this was removed.

Also, remove one unused parameter, simplify regexes and fix a typo.