DRY out JobBase generation
Generate the JobBase parts of both Presubmits and Postsubmit in a single
method.

Additionally, remove some of the regex-handling code for Postsubmits. We
never passed anything else than `true` to the `treatBranchesAsExplicit`
param, so I've removed the parameter and simplified the API. Also, we
took some measures to handle regex-y branches, but the branches come
from **ci-operator** config (not Prowjob config), so they are never
regexes, so we do not need to handle that.