Prowgen: Rewrite pruning stale jobs
We tried to be smart before, but pruning jobs when we are visiting a
file has problems. First, we miss pruning opportunities, second, we
badly treat cases when multiple sources generate jobs into the same
file (like variant jobs).

Add a (default on, but possible to suppress) separate pruning pass after
we generate the jobs. The jobs are labelled with a different label when
we generate them (meaning "we generated this now"). After the generation
pass, we walk everything again and remove the jobs that were not
generated in this run. This also imprves the pruning so that it removes
files with empty jobconfig.