prowgen: temporarily remove upi-src template
This commit is a temporary solution and should be reverted when the
template gets created.  It is not worth it to create a separate
ConfigMap for this type of test.

Tests will fail because of the missing template, but they wouldn't
otherwise be able to be scheduled due to the missing template.  These
tests are supposed to be non-blocking and UPI tests are currently
configured to run 2% of the time, so the impact should be minimal.