Allow users to generate periodic jobs
In addition to generating presubmits and postsubmits, a new field on the
test type allows users to generate periodic jobs. In order to support
generating periodic jobs that require the `src-build` step, we add the
repo under test as the first `extra_ref`. Jobs that do not need to clone
anything will no-op.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>