Use a data structure with more metadata for storing config
When we load every CI Operator configuration, we have the ability to
record a lot of metadata about the files. We should do so and record it
so that it can be used downstream. This patch adds that change and a
conversion function so that the rollout of the change can be
incremental.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>