Add a tool to trigger images postsubmits from changes
When CI Operator configuration is changed in a merge, it may impact the
output images that will be promoted. These changes do not take effect
until the next merge happens to the branch in question, which is often
not close in time to when the configuration changes. Therefore, we add a
component that will run to analyze merged configuration changes and
determine if an image re-build is necessary and trigger it if so.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>