Fix getOwnersHTTP in autoowners
The bug is from the case in which `OWNERS` file DOES exist while
`OWNERS_ALIASES` not. It led to the result of non-existence of
the `OWNERS` file, i.e., `httpResult.ownersFileExists=false`.

The impact of this bug is that we had ignored syncing `OWNERS`
for the repos in the above case.

Test cases have been added to avoid recurrence of this bug in the
feature.