Grab base config from one commit back
When a postsubmit is triggered from a merge, `SHA^1` will refer to the
previous state of the branch before the merge, regardless of how many
commits have been merged in one action. Therefore, this is a safe way
to grab the previous state of the repository. This will work unless
someone manually pushes commits to the repository, which shouldn't
happen, so this is a reasonably resilient, albeit best-effort approach.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>