testgrid-config-generator: Correct some faulty assumptions
The publish steps inside the release config are not good indicators
of ocp or okd. Instead, the naming conventions chosen by the release
should be used.

Fix a bug where the periodics were added to both dashboards by
checking for the correct convention. Also revert the run.sh script
to be correct.