testgrid-config-generator: Update naming and dashboard ordering
Add a new generic informing dashboard for openshift to cover jobs
that fall outside the release spectrum but are still relevant.

Restructure the naming so that we use 'ci' instead of 'okd' (which
was not accurate), and also leave fixes for when we correct the
job naming in prow.

Finally, order the naming so that it is product-release-type-source
and so that the OCP and CI tabs are next to each other - CI is a
leading indicator for OCP, not a separate product.