Add an interactive tool to bootstrap a repo's config
Adding a new repository to the `openshift/release` repository is
nontrivial and requires understanding of three or four different
configuration formats. New component authors rarely care about any of
the nuance on their first pull and will be happy with a smaller set of
options and more defaults. This tool aims to boil down the initial
configuration to a set of simple questions and lets the user edit that
as necessary afterword.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>