Default artifacts directory, set env
When a user runs a test they should be able to set their artifacts
directory to something other than the default, but they should not need
to do anything special to get the default. All good tests create
artifacts so this should not make much of a difference for the amount of
containers we actually run.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>