Add a tool for syncing git content between mirrors
This tools automatically syncs code from which "official" images are
built from their public git repo locations to their mirrors in private
GitHub organization. It is intended to be run in a Prow periodic job
with an interval of less than an hour.

Both source and destination are queried with `git ls-remote`: if they are
already in sync, no further git operations are done. When the destination is
detected to be an empty repo, full `git fetch` is done against a source,
and FETCH_HEAD is then pushed to the destination. When the destination
is a non-empty repository, `git fetch --depth X` is done against the
source and the resulting shallow branch tip is attempted to be pushed.
When this fails, X is exponentially increased and the fetch is retried.
When a threshold is hit, the shallow branch tip is fully fetched with
`--unshallow` and the push is retried again.

Currently the tool does not fail when the destination repository does not exist
at all: this should allow running against full openshift/release while
repository mirrors are created in the private org.