testgrid: Exclude the test container summary junit from testgrid
The test container junit (generated by ci-operator) is already
summarized by the overall test result (if we fail test, we fail
the overall run), but the summary page for test grid will count
it as part of the flake rate and failure percentage. A downside
of this change is that if the test container breaks test grid won't
show a specific red box (but prow would), and also that we can't
see the time the full test suite takes to run in the table graph,
which is an acceptable tradeoff for making the test summary
clearer.