testgrid: Gather more testgrid results for jobs that run less frequently
Many periodics are on 24h or 48h intervals, so test grid's default of
7 days doesn't show much info. For those jobs, use a heuristic to try
and get 100 test results based on the configured interval of the
periodic, between 7d (the default) and 60d (a reasonable max).

Release controller jobs are not changed because they are run at
intervals outside of our control.