ci-secret-bootstrap: simplify censoring formatter setup
Before any secrets are read, the censoring formatter does nothing.
Therefore, setting the formatter to the censoring formatter before we've
read the password and ingested it does nothing of value. We can set the
formatter afterwords and keep the level of complexity here smaller,
removing global variables and other machinery.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>